backend/requirements.txt
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
flask
PyJWT
flask-cors
werkzeug
PyPDF2
python-docx
mysql-connector-python
python-dotenv
gunicorn
reportlab
